feat: Add back and skip button
