+++ nbproject/private/private.xml
@@ -3,6 +3,10 @@
     <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
         <group>
+            <file>file:/E:/FPT/Spring%202024/CSD201/projects/LinkedListFamily/src/Model/Song.java</file>
+            <file>file:/E:/FPT/Spring%202024/CSD201/projects/LinkedListFamily/src/Common/Algorithm.java</file>
+            <file>file:/E:/FPT/Spring%202024/CSD201/projects/LinkedListFamily/src/Model/LinkedList.java</file>
+            <file>file:/E:/FPT/Spring%202024/CSD201/projects/LinkedListFamily/src/Model/Node.java</file>
             <file>file:/E:/FPT/Spring%202024/CSD201/projects/LinkedListFamily/src/Main.java</file>
             <file>file:/E:/FPT/Spring%202024/CSD201/projects/LinkedListFamily/src/Controller/MainController.java</file>
         </group>
