+++ src/Main.java
@@ -7,7 +7,7 @@ import view.Menu;
 public class Main {
     public static void main(String[] args) {
         Node node = new Node();
-        Menu mainMenu = new MainController(node, null);
-        new MainController(node, mainMenu).run();
+        Menu mainMenu = new MainController(null);
+        new MainController(mainMenu).run();
     }
 }
